fix: update search and share tool specs
